Address metadata distribution review feedback
