+++ lightly_studio_view/src/lib/components/DatasetDistributionPanel/HistogramExpandDialog/HistogramExpandDialog.svelte
@@ -1,10 +1,8 @@
 <script lang="ts">
     import * as Dialog from '$lib/components/ui/dialog';
     import { Histogram, type HistogramData, type HistogramRange } from '$lib/components/Histogram';
-    import { Select, type SelectItem } from '$lib/components/Select';
-    import { formatFloat, formatInteger } from '$lib/utils';
-    import { HISTOGRAM_BIN_COUNT_ITEMS } from '../types';
-    import { ValueModeSelect, type ValueMode } from '../PanelHeader/ValueModeSelect';
+    import type { ValueMode } from '../PanelHeader/ValueModeSelect';
+    import HistogramExpandToolbar from './HistogramExpandToolbar/HistogramExpandToolbar.svelte';
 
     interface Props {
         /** Two-way bound flag controlling dialog visibility. */
@@ -45,13 +43,6 @@
     // Measured height of the chart viewport; drives the chart's height budget
     // (bind:clientHeight is backed by a ResizeObserver).
     let chartHeight = $state(0);
-
-    const totalCount = $derived(data.counts.reduce((sum, count) => sum + count, 0));
-
-    const binCountItems: SelectItem[] = HISTOGRAM_BIN_COUNT_ITEMS.map((count) => ({
-        value: String(count),
-        label: `${count} bins`
-    }));
 </script>
 
 <Dialog.Root bind:open>
@@ -62,38 +53,14 @@
                 Click or drag across bars to filter by value range; re-select to reset
             </Dialog.Description>
         </Dialog.Header>
-        <div class="flex flex-wrap items-center justify-between gap-2">
-            <span
-                class="text-xs text-muted-foreground"
-                data-testid="dataset-distribution-expanded-histogram-summary"
-            >
-                {valueMode === 'percentage' ? '100% of ' : ''}{formatInteger(totalCount)}
-                {valueNoun} · {data.counts.length}
-                {data.counts.length === 1 ? 'bin' : 'bins'} · {formatFloat(
-                    data.binEdges[0]
-                )}–{formatFloat(data.binEdges[data.binEdges.length - 1])}
-            </span>
-            <div class="flex items-center gap-1">
-                {#if onValueModeChange}
-                    <ValueModeSelect
-                        value={valueMode}
-                        testId="dataset-distribution-expanded-histogram-value-mode"
-                        onChange={onValueModeChange}
-                    />
-                {/if}
-                {#if onBinCountChange}
-                    <Select
-                        items={binCountItems}
-                        value={String(binCount)}
-                        size="xs"
-                        class="w-28"
-                        testId="dataset-distribution-expanded-bin-count"
-                        selectProps={{ 'aria-label': 'Histogram bin count' }}
-                        onValueChange={(value) => onBinCountChange(Number(value))}
-                    />
-                {/if}
-            </div>
-        </div>
+        <HistogramExpandToolbar
+            {data}
+            {valueNoun}
+            {binCount}
+            {onBinCountChange}
+            {valueMode}
+            {onValueModeChange}
+        />
         <div class="min-h-0 flex-1 dark:[color-scheme:dark]" bind:clientHeight={chartHeight}>
             <Histogram
                 {data}
